# openapi.yaml
openapi: 3.1.0
info:
  title: BugTraceAI CLI API
  description: |
    REST API for the BugTraceAI autonomous security scanning platform.

    BugTraceAI CLI is a headless, AI-powered vulnerability scanner that discovers and validates
    security issues using specialized agents (XSS, SQLi, SSRF, IDOR, LFI, RCE, XXE, JWT, and more).
    This API provides full scan lifecycle management, real-time WebSocket streaming, report generation,
    and runtime configuration.

    **Local beta deployment only.** No authentication is required because access is assumed to be
    limited to the local machine or a trusted LAN. Do not expose the API directly to the Internet.
    ModelLab uses the configured OpenRouter API key and may incur provider costs.
    WebSocket endpoints are available for real-time scan events, findings, and ModelLab progress.

    **Note:** The following endpoints are documented in code but not yet in this spec:
    - POST /api/auth/validate (YAML auth configuration validation)
    - PATCH /api/scans/{scan_id}/auth-config (Update scan authentication)
    - POST /api/tools/totp-verify (TOTP/2FA verification)
  version: 4.0.0-rc1
  contact:
    name: BugTraceAI
    url: https://bugtraceai.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8000
    description: Local development
  - url: http://{host}:8000
    description: Custom deployment
    variables:
      host:
...
paths:
...
  /api/scans:
    post:
      tags: [scans]
      summary: Create scan
      description: |
        Create and start a new security scan against a target URL.
        The scan runs asynchronously — use the status endpoint or WebSocket to track progress.
      operationId: createScan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScanRequest"
      responses:
        "201":
          description: Scan created and started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanStatusResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many concurrent scans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags: [scans]
      summary: List scans
      description: List scan history with pagination and optional status filtering.
      operationId: listScans
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number (1-indexed)
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Items per page
        - name: status_filter
          in: query
          schema:
            $ref: "#/components/schemas/ScanStatus"
          description: Filter by scan status
      responses:
        "200":
          description: Paginated scan list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanListResponse"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/scans/{scan_id}/status:
    get:
      tags: [scans]
      summary: Get scan status
      description: Get current status, progress, and active agent for a scan.
      operationId: getScanStatus
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Current scan status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanStatusResponse"
        "404":
          description: Scan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/scans/{scan_id}/findings:
    get:
      tags: [scans]
      summary: Get scan findings
      description: Get vulnerability findings for a scan with severity/type filtering and pagination.
      operationId: getScanFindings
      parameters:
        - $ref: "#/components/parameters/ScanId"
        - name: severity
          in: query
          schema:
            $ref: "#/components/schemas/Severity"
          description: Filter by severity level
        - name: vuln_type
          in: query
          schema:
            type: string
            example: XSS
          description: "Filter by vulnerability type (e.g. XSS, SQLi, SSRF, IDOR, LFI, RCE, XXE, JWT)"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Paginated findings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingsResponse"
        "400":
          description: Invalid filter or pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Scan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/scans/{scan_id}/stop:
    post:
      tags: [scans]
      summary: Stop scan
      description: Gracefully stop a running scan.
      operationId: stopScan
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Scan stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StopScanResponse"
        "404":
          description: Scan not found or not running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/scans/{scan_id}/pause:
    post:
      tags: [scans]
      summary: Pause scan
      description: Pause a running scan. Can be resumed later.
      operationId: pauseScan
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Scan paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StopScanResponse"
        "404":
          description: Scan not found or not running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/scans/{scan_id}/resume:
    post:
      tags: [scans]
      summary: Resume scan
      description: Resume a previously paused scan.
      operationId: resumeScan
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Scan resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StopScanResponse"
        "404":
          description: Scan not found or not paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/scans/{scan_id}:
    delete:
      tags: [scans]
      summary: Delete scan
      description: |
        Delete a scan and its associated findings and report files.
        Cannot delete a currently running scan.
      operationId: deleteScan
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Scan deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteScanResponse"
        "404":
          description: Scan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Scan is currently running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    ScanId:
      name: scan_id
      in: path
      required: true
      schema:
        type: integer
      description: Unique scan identifier
  schemas:
    ScanStatus:
      type: string
      enum: [PENDING, INITIALIZING, RUNNING, PAUSED, COMPLETED, STOPPED, FAILED]
      description: Scan lifecycle status
    FindingStatus:
      type: string
      enum: [PENDING_VALIDATION, VALIDATED_CONFIRMED, VALIDATED_FALSE_POSITIVE, MANUAL_REVIEW_RECOMMENDED, SKIPPED, ERROR]
      description: Finding validation status
    Severity:
      type: string
      enum: [CRITICAL, HIGH, MEDIUM, LOW, INFO]
      description: Vulnerability severity level
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Human-readable error message
        error_code:
          type: string
          description: Machine-readable error code
...
    CreateScanRequest:
      type: object
      required: [target_url]
      properties:
        target_url:
          type: string
          format: uri
          description: Target URL to scan
          example: https://example.com
        scan_type:
          type: string
          default: full
          description: "Scan type: full, hunter, manager, or focused agent names"
          example: full
        scan_depth:
          type: string
          default: ""
          description: "Exploitation depth: quick, standard, thorough"
          enum: ["", quick, standard, thorough]
        safe_mode:
          type: boolean
          nullable: true
          description: Override global safe mode setting
        max_depth:
          type: integer
          default: 2
          description: Maximum crawl depth
        max_urls:
          type: integer
          default: 20
          description: Maximum URLs to crawl
        resume:
          type: boolean
          default: false
          description: Resume a previous scan
        use_vertical:
          type: boolean
          default: true
          description: Use vertical specialized agents
        focused_agents:
          type: array
          items:
            type: string
          default: []
          description: List of focused agent names
        param:
          type: string
          nullable: true
          description: Specific parameter to target
    ScanStatusResponse:
      type: object
      properties:
        scan_id:
          type: integer
        target:
          type: string
        status:
          $ref: "#/components/schemas/ScanStatus"
        progress:
          type: integer
          minimum: 0
          maximum: 100
        uptime_seconds:
          type: number
          nullable: true
        findings_count:
          type: integer
        active_agent:
          type: string
          nullable: true
        phase:
          type: string
          nullable: true
        origin:
          type: string
          enum: [cli, web, unknown]
          description: Where the scan was launched from
    FindingItem:
      type: object
      properties:
        finding_id:
          type: integer
        type:
          type: string
          description: Vulnerability type (XSS, SQLi, SSRF, etc.)
        severity:
          $ref: "#/components/schemas/Severity"
        details:
          type: string
        payload:
          type: string
          nullable: true
        url:
          type: string
        parameter:
          type: string
          nullable: true
        validated:
          type: boolean
        status:
          $ref: "#/components/schemas/FindingStatus"
        confidence:
          type: number
          nullable: true
          minimum: 0
          maximum: 1
    FindingsResponse:
      type: object
      properties:
        findings:
          type: array
          items:
            $ref: "#/components/schemas/FindingItem"
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        scan_id:
          type: integer
    ScanSummary:
      type: object
      properties:
        scan_id:
          type: integer
        target:
          type: string
        status:
          $ref: "#/components/schemas/ScanStatus"
        progress:
          type: integer
        timestamp:
          type: string
          format: date-time
        origin:
          type: string
          enum: [cli, web, unknown]
        has_report:
          type: boolean
    ScanListResponse:
      type: object
      properties:
        scans:
          type: array
          items:
            $ref: "#/components/schemas/ScanSummary"
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
    StopScanResponse:
      type: object
      properties:
        scan_id:
          type: integer
        status:
          type: string
        message:
          type: string
    DeleteScanResponse:
      type: object
      properties:
        scan_id:
          type: integer
        message:
          type: string
        files_cleaned:
          type: boolean
...